+++ semver-project-gradle-plugin/testFunctional/kotlin/com/javiersc/semver/project/gradle/plugin/PropertiesTest.kt
@@ -3,6 +3,8 @@ package com.javiersc.semver.project.gradle.plugin
 import com.javiersc.gradle.testkit.test.extensions.GradleTestKitTest
 import com.javiersc.kotlin.stdlib.AnsiColor
 import com.javiersc.kotlin.stdlib.ansiColor
+import io.kotest.matchers.string.shouldContain
+import io.kotest.matchers.string.shouldNotContain
 import java.io.File
 import kotlin.test.Test
 
@@ -10,11 +12,14 @@ internal class PropertiesTest : GradleTestKitTest() {
 
     @Test
     fun empty() {
-        runPropertyTestsBasedOnResourceDirectory("empty") {
-            resolve("empty.txt").createNewFile()
-            git.add().addFilepattern(".").call()
-            if (name.endsWith("hash")) git.commit().setMessage("Add empty.txt").call()
-        }
+        runPropertyTestsBasedOnResourceDirectory(
+            resourcePath = "empty",
+            beforeTest = {
+                resolve("empty.txt").createNewFile()
+                git.add().addFilepattern(".").call()
+                if (name.endsWith("hash")) git.commit().setMessage("Add empty.txt").call()
+            },
+        )
     }
 
     @Test
@@ -32,9 +37,38 @@ internal class PropertiesTest : GradleTestKitTest() {
         runPropertyTestsBasedOnResourceDirectory("stage+scope")
     }
 
+    @Test
+    fun `log on all projects`() {
+        gradleTestKitTest("properties/log-on-all-projects") {
+            projectDir.generateInitialCommitAddVersionTagAndAddNewCommit()
+            gradlew("printSemver", "-Psemver.logOnlyOnRootProject=false")
+                .output
+                .shouldContain("semver for sandbox-project: v1.")
+                .shouldContain("semver for library-one: v1.")
+                .shouldContain("semver for library-two: v1.")
+            gradlew("printSemver")
+                .output
+                .shouldContain("semver for sandbox-project: v1.")
+                .shouldContain("semver for library-one: v1.")
+                .shouldContain("semver for library-two: v1.")
+        }
+    }
+
+    @Test
+    fun `log only on root project`() {
+        gradleTestKitTest("properties/log-only-on-root-project") {
+            projectDir.generateInitialCommitAddVersionTagAndAddNewCommit()
+            gradlew("printSemver", "-Psemver.logOnlyOnRootProject=true")
+                .output
+                .shouldContain("semver for sandbox-project: v1.")
+                .shouldNotContain("semver for library-one: v1.")
+                .shouldNotContain("semver for library-two: v1.")
+        }
+    }
+
     private fun runPropertyTestsBasedOnResourceDirectory(
         resourcePath: String,
-        beforeTest: File.() -> Unit = {}
+        beforeTest: File.() -> Unit = {},
     ) {
         val projects =
             getResource("properties/$resourcePath")